requirements.txt
pyserial
tkinter